pyproject.toml
[project]
name = "classical-text-platform"
version = "0.1.0"
description = "고전 텍스트 디지털 서고 플랫폼 — 사람과 LLM이 함께 고전 텍스트를 읽고 번역하고 연구하는 통합 작업 환경"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "hw725" },
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gitpython>=3.1.0",
    "jsonschema>=4.20.0",
    "httpx>=0.27.0",
    "lxml>=5.0.0",
    "fpdf2>=2.8.5",
    "cssselect>=1.4.0",
    "pymupdf>=1.27.1",
    "pillow>=12.1.1",
    "pyyaml>=6.0.3",
    "anthropic>=0.79.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core", "src/parsers", "src/cli", "src/app"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
